Adjust for the comments (#40)
