+++ pyproject.toml
@@ -86,10 +86,6 @@ exclude = [
 # Standard stability settings
 warn_unused_configs = true
 
-[[tool.mypy.overrides]]
-module = "wandb"
-ignore_missing_imports = true
-
 [dependency-groups]
 dev = [
     "pandas-stubs>=3.0.3.260530,<4.0.0",
